refactor(agentspan): extract JsonSchemaTextConverter from AgentCompiler
Move simplifySchema/simplifyNode/inlineRefs/resolveNode/pythonDictRepr —
pure JSON-Schema-to-text converters with no AgentConfig dependency and a
single call site in buildLlmTask — into their own package-private class.
pythonDictRepr becomes private (no caller found outside its own
recursion). No test files touched.
